Extract BranchesSelection to a component
